fastapi/app/client/nest_js_client/models/markdown_res_dto.py
from __future__ import annotations

import datetime
from collections.abc import Mapping
from typing import Any, TypeVar

from attrs import define as _attrs_define
from attrs import field as _attrs_field
from dateutil.parser import isoparse

T = TypeVar("T", bound="MarkdownResDto")


@_attrs_define
class MarkdownResDto:
    """
    Attributes:
        id (float):
        title (str):
        content (str):
        updated_at (datetime.datetime):
        version (float):
    """

    id: float
    title: str
    content: str
    updated_at: datetime.datetime
    version: float
    additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)

    def to_dict(self) -> dict[str, Any]:
        id = self.id

        title = self.title

        content = self.content

        updated_at = self.updated_at.isoformat()

        version = self.version

        field_dict: dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update(
            {
                "id": id,
                "title": title,
                "content": content,
                "updatedAt": updated_at,
                "version": version,
            }
        )

        return field_dict

    @classmethod
    def from_dict(cls: type[T], src_dict: Mapping[str, Any]) -> T:
        d = dict(src_dict)
        id = d.pop("id")

        title = d.pop("title")

        content = d.pop("content")

        updated_at = isoparse(d.pop("updatedAt"))

        version = d.pop("version")

        markdown_res_dto = cls(
            id=id,
            title=title,
            content=content,
            updated_at=updated_at,
            version=version,
        )

        markdown_res_dto.additional_properties = d
        return markdown_res_dto

    @property
    def additional_keys(self) -> list[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties
